Initial responsive fixes, centralized styled components
